dashboard/requirements.txt
streamlit
mysql-connector-python
pandas
plotly
scikit-learn
joblib